pyproject.toml
[project]
name = "grpc-tutor"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "H1rono", email = "hronok66@gmail.com" }
]
dependencies = [
    "grpcio>=1.64.1",
    "grpcio-tools>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[project.scripts]
helloworld-server = "grpc_tutor.greeter_server:serve"
helloworld-client = "grpc_tutor.greeter_client:run"
route-guide-server = "grpc_tutor.route_guide_server:serve"
route-guide-client = "grpc_tutor.route_guide_client:run"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.4.9",
    "types-protobuf>=5.26.0.20240422",
    "mypy>=1.10.0",
    "mypy-protobuf>=3.6.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["py"]

[tool.ruff]
line-length = 90
exclude = ["*_pb2.py", "*_pb2_grpc.py", "*_pb2.pyi", "*_pb2_grpc.pyi"]

[tool.ruff.lint]
# Inherit rules from pycodestyle, pyflakes, flake8-simplify, isort
select = ["F", "E", "SIM", "I"]

[tool.ruff.lint.isort]
combine-as-imports = true
split-on-trailing-comma = true

[tool.mypy]
exclude = [
    '^.*_pb2\.pyi?$',
    '^.*_pb2_grpc\.pyi?$',
]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
